+++ src/main/java/NgoGiaSam/Web_Elaban_be/dao/OrderRespository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.rest.core.annotation.RepositoryRestResource;
 
 @RepositoryRestResource (path = "order")
-public interface OrderRespository extends JpaRepository<Order,Integer> {
+public interface OrderRespository extends JpaRepository<Order,Long> {
 }
